More cleanup + tests are passing locally
