+++ demo/packages/meteor-user-status/client/monitor.js
@@ -1,0 +1,1 @@
+../../../../client/monitor.js
