--- demo/packages/meteor-user-status/monitor.js
@@ -1,1 +1,0 @@
-../../../monitor.js
