+++ demo/packages/meteor-user-status/server/status.js
@@ -1,0 +1,1 @@
+../../../../server/status.js
