--- demo/packages/meteor-user-status/status.js
@@ -1,1 +1,0 @@
-../../../status.js
