+++ package.json
@@ -11,7 +11,8 @@
   "homepage": "https://github.com/Meteor-Community-Packages/meteor-user-status#readme",
   "private": false,
   "scripts": {
-    "lint": "./node_modules/.bin/eslint ."
+    "lint": "./node_modules/.bin/eslint .",
+    "test": "meteor test-packages --once --driver-package test-in-console"
   },
   "devDependencies": {
     "autoprefixer": "^9.6.1",
